+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 
-### Unreleased
+### 0.1.28
+* Added the diffy screenshot:create-baseline command
+* Added the diffy screenshot:set-baseline command
+
+### Released
 
 * Added the diffy project:update command.
 * Add possibility to create multiple projects by a JSON array using project:create command.